+++ projects/core/src/lib/auth/authenticated-user.ts
