+++ projects/core/src/lib/auth/authentication-options.ts
