+++ projects/core/src/lib/auth/authentication-token.interceptor.ts
