+++ projects/core/src/lib/auth/authorization-data.ts
