+++ projects/core/src/lib/auth/login-redirect/login-redirect.component.ts
