+++ projects/core/src/public-api.ts
@@ -1,15 +1,15 @@
 // identityserver
-export { AuthenticationOptions, AUTHENTICATION_OPTIONS } from './lib/identityserver/authentication-options';
-export { AuthenticationService, AuthenticationStateData, AuthenticationEvent, AuthenticationEventType } from './lib/identityserver/authentication.service';
-export { AuthenticatedUser } from './lib/identityserver/authenticated-user';
-export { AuthenticationTokenInterceptor, AUTHENTICATED_REQUEST } from './lib/identityserver/authentication-token.interceptor';
-export { authenticationRoutes } from './lib/identityserver/authentication.routes';
+export { AuthenticationOptions, AUTHENTICATION_OPTIONS } from './lib/auth/authentication-options';
+export { AuthenticationService, AuthenticationStateData, AuthenticationEvent, AuthenticationEventType } from './lib/auth/authentication.service';
+export { AuthenticatedUser } from './lib/auth/authenticated-user';
+export { AuthenticationTokenInterceptor, AUTHENTICATED_REQUEST } from './lib/auth/authentication-token.interceptor';
+export { authenticationRoutes } from './lib/auth/authentication.routes';
 // authorization
-export { AuthorizationService } from './lib/authorization/authorization.service';
-export { AuthorizationContext, AuthorizationPolicy, AUTHORIZATION_POLICY, PermissionAuthorizationPolicy } from './lib/authorization/authorization-policy';
-export { AuthorizationData } from './lib/authorization/authorization-data';
-export { requireAuthorization } from './lib/authorization/authorized.guard';
-export { AuthorizationGuard } from './lib/authorization.guard';
+export { AuthorizationService } from './lib/auth/authorization.service';
+export { AuthorizationContext, AuthorizationPolicy, AUTHORIZATION_POLICY, PermissionAuthorizationPolicy } from './lib/auth/authorization-policy';
+export { AuthorizationData } from './lib/auth/authorization-data';
+export { requireAuthorization } from './lib/auth/authorized.guard';
+export { AuthorizationGuard } from './lib/auth/authorization.guard';
 // logger
 export { LOGGER } from './lib/logger/logger.token';
 export { Logger, BaseLogger, LogLevel, LOG_LEVELS } from './lib/logger/logger';
@@ -26,4 +26,4 @@ export { List, Paged, Search, SearchModel, Auditable, Subject, AuditableModel, S
 export { UnexpectedError } from './lib/errors';
 export { OperationResource, OperationResult, OperationStatus } from './lib/operation-resource';
 
-export { provideAuthentication, OIDC_IFRAME_NAVIGATOR, OIDC_POPUP_NAVIGATOR, OIDC_REDIRECT_NAVIGATOR, OIDC_USER_MANAGER, OIDC_STATE_STORE } from './lib/identityserver/providers';
+export { provideAuth, OIDC_USER_MANAGER } from './lib/auth/providers';
