+++ src/app/app.module.ts
@@ -1,7 +1,7 @@
 import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { APP_INITIALIZER, Injector, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { AuthenticationService, AuthorizationService, provideAuthentication, WINDOW } from '@muziehdesign/angularcore';
+import { AuthenticationService, AuthorizationService, provideAuth, WINDOW } from '@muziehdesign/angularcore';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
@@ -22,7 +22,7 @@ import { AppConfig } from 'src/environments/app-config';
         AppRoutingModule,
     ],
     providers: [
-        provideAuthentication((injector: Injector) => {
+        provideAuth((injector: Injector) => {
             const config = injector.get(AppConfig);
             return config.identity;
         }),
