+++ src/app/core/core.module.ts
@@ -1,6 +1,6 @@
 import { ErrorHandler, NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AUTHORIZATION_POLICY, AuthenticationTokenInterceptor, LOGGER, Logger, PermissionAuthorizationPolicy } from '@muziehdesign/angularcore';
+import { AuthenticationTokenInterceptor, LOGGER, Logger } from '@muziehdesign/angularcore';
 import { GlobalErrorHandler } from './global-error-handler';
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { CustomLogger } from './custom-logger';
@@ -11,7 +11,6 @@ import { CustomLogger } from './custom-logger';
     providers: [
         { provide: LOGGER, useValue: window.console }, // TODO: use actual logger
         { provide: Logger, useClass: CustomLogger },
-        { provide: AUTHORIZATION_POLICY, useClass: PermissionAuthorizationPolicy, multi: true }, 
         { provide: ErrorHandler, useClass: GlobalErrorHandler },
         { provide: HTTP_INTERCEPTORS, useClass: AuthenticationTokenInterceptor, multi: true },
     ],
